+++ lib/src/features/auth/domain/email_input.dart
