+++ lib/src/features/auth/domain/password_input.dart
