+++ lib/src/features/auth/presentation/login/login_form_controller.dart
@@ -1,5 +1,5 @@
-import 'package:betchya_frontend/src/features/auth/models/email_input.dart';
-import 'package:betchya_frontend/src/features/auth/models/password_input.dart';
+import 'package:betchya_frontend/src/features/auth/domain/email_input.dart';
+import 'package:betchya_frontend/src/features/auth/domain/password_input.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:formz/formz.dart';
