+++ lib/src/features/auth/presentation/login/login_screen.dart
@@ -1,5 +1,5 @@
-import 'package:betchya_frontend/src/features/auth/controllers/login_form_controller.dart';
-import 'package:betchya_frontend/src/features/auth/providers/auth_provider.dart';
+import 'package:betchya_frontend/src/features/auth/presentation/login/login_form_controller.dart';
+import 'package:betchya_frontend/src/features/auth/presentation/auth_provider.dart';
 import 'package:betchya_frontend/src/routing/app_router.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
