+++ lib/src/routing/app_router.dart
@@ -1,7 +1,7 @@
-import 'package:betchya_frontend/src/features/auth/presentation/home_screen.dart';
-import 'package:betchya_frontend/src/features/auth/presentation/login_screen.dart';
-import 'package:betchya_frontend/src/features/auth/presentation/signup_screen.dart';
-import 'package:betchya_frontend/src/features/auth/providers/auth_provider.dart';
+import 'package:betchya_frontend/src/features/home/presentation/home_screen.dart';
+import 'package:betchya_frontend/src/features/auth/presentation/login/login_screen.dart';
+import 'package:betchya_frontend/src/features/auth/presentation/signup/signup_screen.dart';
+import 'package:betchya_frontend/src/features/auth/presentation/auth_provider.dart';
 import 'package:betchya_frontend/src/routing/go_router_refresh_stream.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
