--- test/helpers/pump_app.dart
@@ -1,36 +1,0 @@
-import 'package:betchya_frontend/src/l10n/l10n.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-extension PumpApp on WidgetTester {
-  Future<void> pumpApp(
-    Widget widget, {
-    List<Override> overrides = const [],
-  }) {
-    return pumpWidget(
-      ProviderScope(
-        overrides: overrides,
-        child: MaterialApp(
-          localizationsDelegates: AppLocalizations.localizationsDelegates,
-          supportedLocales: AppLocalizations.supportedLocales,
-          home: widget,
-        ),
-      ),
-    );
-  }
-}
-
-/// Pumps a widget with the given provider overrides
-Future<void> pumpWidgetWithOverrides({
-  required WidgetTester tester,
-  required List<Override> overrides,
-  required Widget child,
-}) async {
-  await tester.pumpWidget(
-    ProviderScope(
-      overrides: overrides,
-      child: MaterialApp(home: child),
-    ),
-  );
-}
