+++ test/src/features/auth/presentation/robots/login_screen_robot.dart
@@ -1,4 +1,4 @@
-import 'package:betchya_frontend/src/features/auth/presentation/home_screen.dart';
+import 'package:betchya_frontend/src/features/home/presentation/home_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
