+++ test/src/features/home/presentation/home_screen_test.dart
@@ -1,5 +1,5 @@
-import 'package:betchya_frontend/src/features/auth/presentation/home_screen.dart';
-import 'package:betchya_frontend/src/features/auth/providers/auth_provider.dart';
+import 'package:betchya_frontend/src/features/home/presentation/home_screen.dart';
+import 'package:betchya_frontend/src/features/auth/presentation/auth_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
